requirements.txt
Flask
azure-cognitiveservices-speech
requests
email
